actor-tourism-it/requirements.txt
apify==2.7.3
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.3.0
psycopg2-binary>=2.9.9
